feat(ui): eliminate StudioContext and consolidate state in Studio container
