+++ web/src/components/LandingView.stories.tsx
@@ -1,6 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { LandingView } from './LandingView';
-import { MockStudioProvider } from '../context/MockStudioProvider';
 
 const meta: Meta<typeof LandingView> = {
   title: 'Components/LandingView',
@@ -10,10 +9,4 @@ const meta: Meta<typeof LandingView> = {
 export default meta;
 type Story = StoryObj<typeof LandingView>;
 
-export const Default: Story = {
-  decorators: [
-    (Story) => (
-      <MockStudioProvider mockValues={{}}>{Story()}</MockStudioProvider>
-    ),
-  ],
-};
+export const Default: Story = {};
