--- web/tests/fixtures/studioContext.ts
@@ -1,44 +1,0 @@
-import { vi } from 'vitest';
-import { type StudioContextType } from '../../src/context/StudioContext';
-
-export const createMockStudioContext = (
-  overrides: Partial<StudioContextType> = {}
-): StudioContextType => ({
-  svgContent: null,
-  setSvgContent: vi.fn(),
-  fileName: 'test.mp4',
-  setFileName: vi.fn(),
-  duration: 5,
-  setDuration: vi.fn(),
-  hold: 0,
-  setHold: vi.fn(),
-  fps: 60,
-  setFps: vi.fn(),
-  preset: '1080p',
-  setPreset: vi.fn(),
-  scale: 1,
-  setScale: vi.fn(),
-  backgroundColor: '#ffffff',
-  setBackgroundColor: vi.fn(),
-  format: 'mp4',
-  setFormat: vi.fn(),
-  isTransparent: false,
-  setIsTransparent: vi.fn(),
-  captureMethod: 'optimal',
-  setCaptureMethod: vi.fn(),
-  isDragging: false,
-  setIsDragging: vi.fn(),
-  renderedUrl: null,
-  setRenderedUrl: vi.fn(),
-  fileSize: null,
-  setFileSize: vi.fn(),
-  originalDim: { width: 500, height: 500, isDimensionsDetected: true },
-  targetDim: { width: 500, height: 500 },
-  state: { isRendering: false, status: 'Idle', progress: 0 },
-  handleStartRender: vi.fn(),
-  cancel: vi.fn(),
-  clearError: vi.fn(),
-  downloadResult: vi.fn(),
-  rendererRef: { current: null },
-  ...overrides,
-});
